+++ docs/app/components/content/examples/slideover/SlideoverFooterSlotExample.vue
@@ -3,7 +3,12 @@ const open = ref(false)
 </script>
 
 <template>
-  <B24Slideover v-model:open="open" title="Slideover with footer" description="This is useful when you want a form in a Slideover." :b24ui="{ footer: 'justify-end' }">
+  <B24Slideover
+    v-model:open="open"
+    title="Slideover with footer"
+    description="This is useful when you want a form in a Slideover."
+    :b24ui="{ footer: 'justify-between' }"
+  >
     <B24Button label="Open" />
 
     <template #body>
@@ -11,8 +16,16 @@ const open = ref(false)
     </template>
 
     <template #footer="{ close }">
-      <B24Button label="Cancel" @click="close" />
-      <B24Button label="Submit" color="air-primary-success" />
+      <div class="w-1/5 flex justify-start" />
+      <div class="w-full flex flex-row justify-center gap-[10px]">
+        <B24ModalDialogClose>
+          <B24Button label="Send" color="air-primary" />
+        </B24ModalDialogClose>
+        <B24Button label="Cancel" color="air-tertiary" @click="close" />
+      </div>
+      <div class="w-1/5 flex justify-end">
+        <B24Button label="Full version" size="sm" color="air-tertiary-no-accent" />
+      </div>
     </template>
   </B24Slideover>
 </template>