+++ docs/app/components/content/examples/slideover/SlideoverOpenExample.vue
@@ -7,11 +7,11 @@ defineShortcuts({
 </script>
 
 <template>
-  <B24Slideover v-model:open="open">
+  <B24Slideover v-model:open="open" :b24ui="{ content: 'light bg-(--ui-color-base-7) sm:top-[300px] sm:max-h-[calc(100%-300px)]' }">
     <B24Button label="Open" />
 
     <template #content>
-      <Placeholder class="h-full m-4" />
+      <Placeholder class="edge-light h-full m-4" />
     </template>
   </B24Slideover>
 </template>